Список бюджетов вынесен в компонент
